+++ databallpy/__init__.py
@@ -4,11 +4,8 @@ from databallpy.utils.get_game import (
     Game,
     get_game,
     get_game_from_kloppy,
-    get_match,
     get_open_game,
-    get_open_match,
     get_saved_game,
-    get_saved_match,
 )
 from databallpy.utils.logging import create_logger
 from databallpy.utils.to_xml import Event, LabelDict, events_to_xml