+++ databallpy/features/__init__.py
@@ -1,13 +1,6 @@
 from databallpy.features.angle import get_smallest_angle
-from databallpy.features.covered_distance import get_covered_distance
-from databallpy.features.differentiate import add_acceleration, add_velocity
 from databallpy.features.pitch_control import (
-    get_approximate_voronoi,
-    get_pitch_control,
     get_pitch_control_single_frame,
     get_player_influence,
     get_team_influence,
 )
-from databallpy.features.player_possession import get_individual_player_possession
-from databallpy.features.pressure import get_pressure_on_player
-from databallpy.features.team_possession import add_team_possession