+++ tests/features/test_pressure.py
@@ -1,4 +1,3 @@
-import math
 import unittest
 
 import numpy as np
@@ -9,7 +8,6 @@ from databallpy.features.pressure import (
     calculate_l,
     calculate_variable_dfront,
     calculate_z,
-    get_pressure_on_player,
 )
 
 
@@ -27,53 +25,6 @@ class TestPressure(unittest.TestCase):
                 "away_3_y": 30.0,
             }
         )
-
-    def test_get_pressure_on_player_specified_d_front(self):
-        d_front = calculate_variable_dfront(
-            self.td_frame, "home_1", max_d_front=9, pitch_length=100.0
-        )
-        z1 = calculate_z(self.td_frame, "home_1", "away_1", pitch_length=100.0)
-        z2 = calculate_z(self.td_frame, "home_1", "away_2", pitch_length=100.0)
-        l1 = calculate_l(d_back=3.0, d_front=d_front, z=z1)
-        l2 = calculate_l(d_back=3.0, d_front=d_front, z=z2)
-        dist1 = math.dist([1, 1], [2, 2])
-        dist2 = math.dist([1, 1], [3, 3])
-
-        pres1 = (1 - dist1 / l1) ** 1.75 * 100
-        pres2 = (1 - dist2 / l2) ** 1.75 * 100
-        expected_pressure = pres1 + pres2
-
-        result = get_pressure_on_player(
-            self.td_frame,
-            "home_1",
-            pitch_size=[100.0, 50.0],
-            d_front="variable",
-            d_back=3.0,
-            q=1.75,
-        )
-        self.assertAlmostEqual(result, expected_pressure, places=4)
-
-    def test_get_pressure_on_player_variable_d_front(self):
-        z1 = calculate_z(self.td_frame, "home_1", "away_1", pitch_length=100.0)
-        z2 = calculate_z(self.td_frame, "home_1", "away_2", pitch_length=100.0)
-        l1 = calculate_l(d_back=3.0, d_front=9.0, z=z1)
-        l2 = calculate_l(d_back=3.0, d_front=9.0, z=z2)
-        dist1 = math.dist([1, 1], [2, 2])
-        dist2 = math.dist([1, 1], [3, 3])
-
-        pres1 = (1 - dist1 / l1) ** 1.75 * 100
-        pres2 = (1 - dist2 / l2) ** 1.75 * 100
-        expected_pressure = pres1 + pres2
-
-        result = get_pressure_on_player(
-            self.td_frame,
-            "home_1",
-            pitch_size=[100.0, 50.0],
-            d_front=9.0,
-            d_back=3.0,
-            q=1.75,
-        )
-        self.assertAlmostEqual(result, expected_pressure, places=4)
 
     def test_calculate_variable_dfront(self):
         expected = 7.4505
@@ -96,14 +47,3 @@ class TestPressure(unittest.TestCase):
         expected = 3.1046
         res = calculate_l(d_back=3.0, d_front=5.0, z=0.2)
         self.assertAlmostEqual(res, expected, places=4)
-
-    def test_get_pressure_on_player_wrong_input(self):
-        with self.assertRaises(TypeError):
-            get_pressure_on_player(
-                "td_frame",
-                "home_1",
-                pitch_size=[100.0, 50.0],
-                d_front="variable",
-                d_back=3.0,
-                q=1.75,
-            )